test: move e2e case to stable environment
