+++ .github/workflows/tests-e2e.yaml
@@ -24,12 +24,6 @@ jobs:
       - name: Install local dependencies
         run: npm ci
 
-      - name: Install global dependencies
-        run: npm install --global bee-dev
-
-      - name: Run bee-dev
-        run: bee-dev --port 1633 &
-
       - name: Tests
         run: npm run test.e2e
